perf: render input and output
