+++ packages/global/core/module/template/system/http.ts
@@ -1,8 +1,8 @@
 import { FlowNodeInputTypeEnum, FlowNodeTypeEnum } from '../../node/constant';
 import { FlowModuleTemplateType } from '../../type';
 import { ModuleDataTypeEnum, ModuleInputKeyEnum, ModuleTemplateTypeEnum } from '../../constants';
-import { Input_Template_TFSwitch } from '../input';
-import { Output_Template_Finish } from '../output';
+import { Input_Template_AddInputParam, Input_Template_TFSwitch } from '../input';
+import { Output_Template_AddOutput, Output_Template_Finish } from '../output';
 
 export const HttpModule: FlowModuleTemplateType = {
   id: FlowNodeTypeEnum.httpRequest,
@@ -55,7 +55,8 @@ export const HttpModule: FlowModuleTemplateType = {
       required: false,
       showTargetInApp: false,
       showTargetInPlugin: false
-    }
+    },
+    Input_Template_AddInputParam
   ],
-  outputs: [Output_Template_Finish]
+  outputs: [Output_Template_Finish, Output_Template_AddOutput]
 };
