+++ projects/app/src/components/core/module/Flow/components/nodes/NodeTextEditor.tsx
@@ -5,12 +5,7 @@ import { FlowModuleItemType } from '@fastgpt/global/core/module/type.d';
 import Divider from '../modules/Divider';
 import Container from '../modules/Container';
 import RenderInput from '../render/RenderInput';
-import { Button } from '@chakra-ui/react';
-import { SmallAddIcon } from '@chakra-ui/icons';
-import { FlowNodeInputTypeEnum } from '@fastgpt/global/core/module/node/constant';
-import { ModuleDataTypeEnum, ModuleInputKeyEnum } from '@fastgpt/global/core/module/constants';
-import { customAlphabet } from 'nanoid';
-const nanoid = customAlphabet('abcdefghijklmnopqrstuvwxyz1234567890', 6);
+import { ModuleInputKeyEnum } from '@fastgpt/global/core/module/constants';
 import { onChangeNode } from '../../FlowProvider';
 import RenderOutput from '../render/RenderOutput';
 import { useTranslation } from 'next-i18next';
@@ -52,28 +47,6 @@ const NodeTextEditor = ({ data }: NodeProps<FlowModuleItemType>) => {
             }
           }}
         />
-        <Button
-          variant={'base'}
-          mt={5}
-          leftIcon={<SmallAddIcon />}
-          onClick={() => {
-            const key = nanoid();
-            onChangeNode({
-              moduleId,
-              type: 'addInput',
-              key,
-              value: {
-                key,
-                valueType: ModuleDataTypeEnum.string,
-                type: FlowNodeInputTypeEnum.target,
-                label: `入参${inputs.length - 1}`,
-                edit: true
-              }
-            });
-          }}
-        >
-          {t('core.module.input.Add Input')}
-        </Button>
       </Container>
       <Divider text="Output" />
       <Container>
