+++ projects/app/src/components/core/module/Flow/components/render/RenderInput/type.d.ts
@@ -1,0 +1,7 @@
+import { FlowNodeInputItemType } from '@fastgpt/global/core/module/node/type';
+
+export type RenderInputProps = {
+  inputs?: FlowNodeInputItemType[];
+  item: FlowNodeInputItemType;
+  moduleId: string;
+};
