+++ projects/app/src/components/core/module/Flow/components/render/RenderOutput/type.d.ts
@@ -1,0 +1,7 @@
+import { FlowNodeOutputItemType } from '@fastgpt/global/core/module/node/type';
+
+export type RenderOutputProps = {
+  outputs?: FlowNodeOutputItemType[];
+  item: FlowNodeOutputItemType;
+  moduleId: string;
+};
